feat: add keybindings
