+++ model.go
@@ -1,15 +1,15 @@
 package boa
 
 import (
-	"strings"
-	"unicode"
-
+	"github.com/charmbracelet/bubbles/key"
 	"github.com/charmbracelet/bubbles/list"
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 	"github.com/spf13/cobra"
 )
+
+const spacebar = " "
 
 // cmdModel Implements tea.Model. It provides an interactive
 // help and usage tui component for bubbletea programs.
@@ -29,10 +29,12 @@ func newCmdModel(cmd *cobra.Command) *cmdModel {
 	subCmds := getSubCommands(cmd)
 	l := newSubCmdsList(subCmds)
 	return &cmdModel{
-		cmd:           cmd,
-		subCmds:       subCmds,
-		list:          l,
-		viewport:      &viewport.Model{},
+		cmd:     cmd,
+		subCmds: subCmds,
+		list:    l,
+		viewport: &viewport.Model{
+			KeyMap: viewPortKeyMap(),
+		},
 		contentHeight: lipgloss.Height(usage(cmd, l)),
 	}
 }
@@ -76,7 +78,7 @@ func (m cmdModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				m.viewport.SetYOffset(0)
 			}
 			return m, nil
-		case "backspace":
+		case "b":
 			if m.cmd.HasParent() {
 				m.cmd = m.cmd.Parent()
 				subCmds := getSubCommands(m.cmd)
@@ -98,69 +100,12 @@ func (m cmdModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	newViewport, viewPortCmd := m.viewport.Update(msg)
 	// point to new viewport
 	m.viewport = &newViewport
+	m.viewport.KeyMap = viewPortKeyMap()
 	if m.viewport.Height > m.contentHeight+lipgloss.Height(footer(m.contentHeight, m.windowHeight)) {
 		m.viewport.Height = m.contentHeight
 	}
 	cmds = append(cmds, listCmd, viewPortCmd)
 	return m, tea.Batch(cmds...)
-}
-
-func usage(cmd *cobra.Command, list list.Model) string {
-	usageText := strings.Builder{}
-
-	cmdTitle := ""
-	if !cmd.HasParent() {
-		rootCmdName := SectionStyle.Render(cmd.Root().Name() + " " + cmd.Root().Version)
-		rootCmdLong := SubTitleSytle.Render(cmd.Root().Long)
-		cmdTitle = TitleStyle.Foreground(lipgloss.AdaptiveColor{Light: darkGrey, Dark: white}).
-			Render(lipgloss.JoinVertical(lipgloss.Top, rootCmdName, rootCmdLong))
-	}
-	usageText.WriteString(cmdTitle + "\n")
-
-	cmdSection := SectionStyle.Render("Cmd Description:")
-	short := TextStyle.Render(cmd.Short)
-
-	usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, cmdSection, short) + "\n")
-
-	if cmd.Runnable() {
-		usage := SectionStyle.Render("Usage:")
-		useLine := TextStyle.Render(cmd.UseLine())
-		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, usage, useLine) + "\n")
-		if cmd.HasAvailableSubCommands() {
-			commandPath := TextStyle.Render(cmd.CommandPath() + " [command]")
-			usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, commandPath) + "\n")
-		}
-	}
-
-	if len(cmd.Aliases) > 0 {
-		aliases := SectionStyle.Render("Aliases:")
-		nameAndAlias := TextStyle.Render(cmd.NameAndAliases())
-		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, aliases, nameAndAlias) + "\n")
-	}
-
-	if cmd.HasAvailableLocalFlags() {
-		localFlags := SectionStyle.Render("Flags:")
-		flagUsage := TextStyle.Render(strings.TrimRightFunc(cmd.LocalFlags().FlagUsages(), unicode.IsSpace))
-		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, localFlags, flagUsage) + "\n")
-	}
-	if cmd.HasAvailableInheritedFlags() {
-		globalFlags := SectionStyle.Render("Global Flags:")
-		flagUsage := TextStyle.Render(strings.TrimRightFunc(cmd.InheritedFlags().FlagUsages(), unicode.IsSpace))
-		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, globalFlags, flagUsage) + "\n")
-	}
-
-	if cmd.HasExample() {
-		examples := SectionStyle.Render("Examples:")
-		example := TextStyle.Render(cmd.Example)
-		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, examples, example) + "\n")
-	}
-
-	if cmd.HasAvailableSubCommands() {
-		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, list.View()))
-	}
-
-	usageCard := BorderStyle.Render(usageText.String() + "\n")
-	return usageCard
 }
 
 // View renders the program's UI, which is just a string.
@@ -169,8 +114,25 @@ func (m cmdModel) View() string {
 	return lipgloss.JoinVertical(lipgloss.Top, m.viewport.View(), footer(m.contentHeight, m.windowHeight))
 }
 
-func footer(contentHeight int, windowHeight int) string {
-	help := InfoStyle.Render("↑/k up • ↓/j down • / to filter • backspace to go back • enter to select • q, ctrl+c to quit")
-	scroll := InfoStyle.Render("use mouse scroll or space to see full usage")
-	return lipgloss.JoinVertical(lipgloss.Top, help, scroll)
+func viewPortKeyMap() viewport.KeyMap {
+	return viewport.KeyMap{
+		PageDown: key.NewBinding(
+			key.WithKeys("pgdown", spacebar, "f"),
+		),
+		PageUp: key.NewBinding(
+			key.WithKeys("pgup", "b"),
+		),
+		HalfPageUp: key.NewBinding(
+			key.WithKeys("u", "ctrl+u"),
+		),
+		HalfPageDown: key.NewBinding(
+			key.WithKeys("d", "ctrl+d"),
+		),
+		Up: key.NewBinding(
+			key.WithKeys("ctrl+k"),
+		),
+		Down: key.NewBinding(
+			key.WithKeys("ctrl+j"),
+		),
+	}
 }
