+++ styles.go
@@ -1,7 +1,12 @@
 package boa
 
 import (
+	"strings"
+	"unicode"
+
+	"github.com/charmbracelet/bubbles/list"
 	"github.com/charmbracelet/lipgloss"
+	"github.com/spf13/cobra"
 )
 
 const (
@@ -54,3 +59,70 @@ var (
 	InfoStyle = lipgloss.NewStyle().Bold(true).Width(width).Align(lipgloss.Center).
 			Foreground(lipgloss.AdaptiveColor{Light: darkGrey, Dark: white})
 )
+
+func usage(cmd *cobra.Command, list list.Model) string {
+	usageText := strings.Builder{}
+
+	cmdTitle := ""
+	if !cmd.HasParent() {
+		rootCmdName := SectionStyle.Render(cmd.Root().Name() + " " + cmd.Root().Version)
+		rootCmdLong := SubTitleSytle.Render(cmd.Root().Long)
+		cmdTitle = TitleStyle.Foreground(lipgloss.AdaptiveColor{Light: darkGrey, Dark: white}).
+			Render(lipgloss.JoinVertical(lipgloss.Top, rootCmdName, rootCmdLong))
+	}
+	usageText.WriteString(cmdTitle + "\n")
+
+	cmdSection := SectionStyle.Render("Cmd Description:")
+	short := TextStyle.Render(cmd.Short)
+
+	usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, cmdSection, short) + "\n")
+
+	if cmd.Runnable() {
+		usage := SectionStyle.Render("Usage:")
+		useLine := TextStyle.Render(cmd.UseLine())
+		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, usage, useLine) + "\n")
+		if cmd.HasAvailableSubCommands() {
+			commandPath := TextStyle.Render(cmd.CommandPath() + " [command]")
+			usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, commandPath) + "\n")
+		}
+	}
+
+	if len(cmd.Aliases) > 0 {
+		aliases := SectionStyle.Render("Aliases:")
+		nameAndAlias := TextStyle.Render(cmd.NameAndAliases())
+		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, aliases, nameAndAlias) + "\n")
+	}
+
+	if cmd.HasAvailableLocalFlags() {
+		localFlags := SectionStyle.Render("Flags:")
+		flagUsage := TextStyle.Render(strings.TrimRightFunc(cmd.LocalFlags().FlagUsages(), unicode.IsSpace))
+		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, localFlags, flagUsage) + "\n")
+	}
+	if cmd.HasAvailableInheritedFlags() {
+		globalFlags := SectionStyle.Render("Global Flags:")
+		flagUsage := TextStyle.Render(strings.TrimRightFunc(cmd.InheritedFlags().FlagUsages(), unicode.IsSpace))
+		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, globalFlags, flagUsage) + "\n")
+	}
+
+	if cmd.HasExample() {
+		examples := SectionStyle.Render("Examples:")
+		example := TextStyle.Render(cmd.Example)
+		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, examples, example) + "\n")
+	}
+
+	if cmd.HasAvailableSubCommands() {
+		usageText.WriteString(lipgloss.JoinVertical(lipgloss.Top, list.View()))
+	}
+
+	usageCard := BorderStyle.Render(usageText.String() + "\n")
+	return usageCard
+}
+
+func footer(contentHeight int, windowHeight int) string {
+	var help, scroll string
+	help = InfoStyle.Render("↑/k up • ↓/j down • / to filter • b to go back • enter to select • q, ctrl+c to quit")
+	if contentHeight > windowHeight {
+		scroll = InfoStyle.Render("ctrl+k up • ctrl+j down • mouse to scroll")
+	}
+	return lipgloss.JoinVertical(lipgloss.Top, help, scroll)
+}
